Remove slang-win-stream
